pyproject.toml
[tool.poetry]
name = "human_val"
version = "0.1.0"
description = ""
authors = ["Jorge Ruiz <jorge.ruiz@itcl.es>"]
readme = "README.md"

packages = [{include = "human_val"}]
include = [{path = "models/onnx", format = "wheel"}]
exclude = ["human_val/**/tests*"]


[tool.poetry.dependencies]
python = ">=3.6,<3.9"
numpy = "1.18"
onnxruntime-gpu = {version = "1.9"}
pillow = ">=8.0"
attrs = "22.2"

#onnxruntime-gpu-jetson = {url="https://nvidia.box.com/shared/static/w3dezb26wog78rwm2yf2yhh578r5l144.whl"}



[tool.poetry.group.dev.dependencies]
pytest = ">=6"


[build-system]
requires = ["poetry-core>=1.1.0a6"]
build-backend = "poetry.core.masonry.api"
